Add Traffic Ops get_parameters script, postinstall
This adds the get_parameters script, which gets profiles from an
existing Traffic Ops and replaces CDN-specific with template values.

This also adds a set of profiles for postinstall, and changes the
postinstall script to read the profiles, set the template values to
its user-specified parameters, and import them into Traffic Ops.