Remove TO clone ds unnecessary profile check
This removes, from the 'Clone DS assignments' feature,  the SQL query
for profiles in the same CDN, and filtering for only those profiles.

This does not change functionality, since we now filter later for
servers in the same CDN, and since the `server` table `profile`
column is both NOT NULL and a FOREIGN KEY into the `profile` table
(so servers must have profiles, i.e. we'll never now get a server
without a profile that we wouldn't have got before, since that isn't
possible).