Change TM2 DataRequest to be func, not go+chan
Changes `http_server` to take a function, instead of a channel for a
manager goroutine. Since threadsafe vars are now used, there's no
reason for a manager goroutine, and a lambda is much faster than
locking for a chan send, running another goroutine, and locking again
for a chan return.

Fixes #1831