Fix TM2 DS kbps logic to be per-cache
Fixes delivery service kbps calculations to be done per-cache. This
fixes the kbps calculation and last total bytes, when a cache goes
down. Formerly, when a cache went down, that cache's bytes were in the
old total, and not the new, and resulted in wrong numbers, including
potentially negative kbps.

Issue #1977