Fix TM2 DsStats IsHealthy to match IsAvailable
Fixes Traffic Monitor 2.0 `/publish/DsStats` endpoint `isHealthy` stat
to match TM1 and be true iff `isAvailable` is true. The isHealthy stat
was supposed to represent whether the delivery service is healthy,
as defined by the maximum kbps (and other data?) defined in Traffic
Ops. It may or may not need fixed to reflect this in the future.