Add TM2 config for health-to-stat ratio.
Adds Traffic Monitor 2.0 config setting for health-to-stat ration.
This is only used if a 'health' poll interval exists in Traffic Ops
but no 'stat' interval, which is the case for Traffic Monitor 2.0.

If both 'health' and 'stat' exist, they will be used appropriately
and this config setting will be ignored. If only 'health' exists, it
will be used for 'stat' (which is the big query, like the only query
in Traffic Monitor 1.0), and 'health' will be the given ratio. For
example, to set the health poll at a quarter the stat poll, set the
config value to 4. So, a stat poll of 8 becomes 8/4=2 seconds. To use
the same value for both, set the config to 1.

This only exists to make it possible to poll health at a faster
rate in an existing Traffic Control deployment with no parameter
changes. Once deployed, the parameters should be set appropriately
and this config will not be used.