Fix statHistory to be newest-first
Fixes statHistory to be newest-first, and fixes readers to expect the
same. It was oldest-first, with inconsistent reader expectations.

Fixes total bandwidth to include mids, which appears to match TM 1.0.

Fixes a result error check to check the real Error, instead of
PrecomputedData's Errors, and fixes the Result PrecomptuedData
member to not be anonymous, requiring explicit access, to prevent
this kind of accident again.

Fixes srvhttp to include the proper content type header for files.