Add TM2 InsomniacPoll to workaround Sleep bug
Adds Traffic Monitor 2.0 InsomniacPoll, and a config setting for it,
which uses a single poller goroutine which busywaits and dispatches
pollers when necessary, and never sleeps. The regular Sleep Poll
runs a goroutine per poll, each of which sleeps for the poll interval.

This works around an issue wherein Sleep gets progressively slower.
E.g. a 8s poll takes 9s after a day, and 20s after a week. This issue
has been observed on OpenStack with CentOS 6.5 and kernel 2.6.32.
Users who don't experience the Sleep issue should definitely use
the original Sleep poll (via config setting http_poll_no_sleep),
as it's more efficient, idiomatic, faster, and safer. If and when
the Sleep issue is diagnosed and fixed, the InsomniacPoll should
be removed.