Reorganized .rat-excludes.
There were several large directories excluded that contained code
that needed to be analyzed. I re-did the list of excludes to reduce
the chances that future updates that require analysis will escape it
by virtue of accidentally matching an exclusion rule.

Additionally, I included a "comment" regarding each exclusion line,
such as the regex allows. This makes it much clearer why various
files are being excluded and encourages a bit of organization in
the file.