Fix TM2 global variable to be a safe local
This wasn't technically a thread safety issue, because the integer was
only ever checked as a boolean. But it's a dangerous precedent, as in
most cases this design pattern would be unsafe. Changing it to a local
variable, and making the function 'pure', is generally safer, easier
to understand, and doesn't give anyone the idea that pattern is okay.