Add global build script.
This splits out the somewhat messy docker build command into its own script and adds a global build script.
There’s still work to be done to clean up docker build scripts, but this provides a basic way to build top to bottom
In a single, simple command.

This is a step toward addressing TC-180, but falls short of cleaning up everything that requires cleaning up.