Add TM2 getting CDN name from Traffic Ops
Adds Traffic Monitor 2.0 getting the name of the CDN to monitor from
Traffic Ops. The CDN is the CDN that this Traffic Monitor is assigned
to in TO (known by the hostname command). If this TO hostname doesn't
exist in Traffic Ops, an error is logged and the old traffic_ops.cfg
cdnName value is used. If a config cdnName is set which differs from
the TO CDN, the TO CDN is used, and a warning is logged (TO is the
Source of Truth).

This makes deployment easier and safer: Traffic Monitor deployment
configs need only the Traffic Ops URL and login, and the Traffic Ops
CDN will be used, so it isn't possible to accidentally deploy a
Monitor configured to use a different CDN than is in Traffic Ops.