Allow specification of the 'id' parameter for storage.config
This patch adds three new parameters to control/configure the seed
strings for the consistent hash used for cachekey->drive mapping in
ATS... This allows an arbitrary string to be assigned as the key for
drives instead of relying on the OS's device node name as the seed for
the hash. The three parameters added are:
 * Drive_Hash_Id
 * RAM_Drive_Hash_Id
 * SSD_Drive_Hash_Id
And they specify the id keys for spinning, ram, and ssd devices.

If this parameter is missing from the TO config for a server, the
generated file ommits any specification of the id parameter (preserving
the previous file generation behavior).

Also note that the csv id strings for these three new parameters should
order their entries to match the order of the Drive_Letters parameter.