WIP Add origin request limiting
This adds origin protection, limiting requests per key to 1. Currently
requests are limited to 1 per key (Method+URL).

This is a Work In Progress. Currently, it will only work for content
which is cacheable. If no-cache content is sent over the proxy, it
will fall over, since only 1 request will be allowed thru at once.

This also doesn't protect multiple keys. If a million requests are
made to a million different URLs (keys), the origin will still get
hit. That must also be protected against.