[TC-287] Add support for per-DeliveryService routing names
Add a routing_name column to the DeliveryService table and populate it
to 'tr' and 'edge' for HTTP and DNS delivery services, respectively.

To make upgrading seamless for CDNs using a custom http.routing.name
(configured in the TR), provide a temporary upgrade parameter to
populate the routing name for HTTP Delivery Services. Otherwise HTTP
DSes will get routing_name = "tr".

Replace all hardcoded/assumed routing names of 'edge' or 'tr' in Traffic
Ops with the proper routing name configured for that delivery service.

Update the API/UI to support custom routing names for delivery services.
To prevent breaking the API, "routingName" is an optional field (the DB
schema provides an arbitrary default).