Remove failing "migratedb" step from traffic_ops upgrade
The migratedb script has a dependency on the goose binary and hardcodes
the PATH environment variable within the script itself. This hardcoded
PATH doesn't include the directory where the goose binary is installed
to via postinstall. If goose was installed outside of postinstall, it's
unlikely that the binary would be found in the hardcoded PATH.