Fix TM CRConfig for empty string numbers
Changes the CRConfig struct to not use `,string`, which fails to parse
when the value is "".

This is a critical issue which causes the TM to fail to start, or
while running, fail to get new CRConfigs, if a value is "", for
example, if a delivery service has its TTL empty.

Fixes #1116