Fix listener InterceptConn bytes
Fixes the ConnMap to properly remove and add connections when the
Conn changes to Idle/Closed and Active.

The original code assumed Conn was a high-level abstraction unique
to each HTTP Handler. It is not; Conns are re-used. This fixes
them to be removed from the ConnMap when the Conn switches to Idle,
and added back when an Idle Conn switches to Active.

This also importantly zeros the byte counts when the Conn switches
to Idle. This is important, because the Go language calls the
Callback after some, possibly all, bytes have been read. In order to
accurately get Read counts, bytes must be zeroed and begin counting
again upon the callback for Idle, not Active.