Fix how deep NetworkNodes are loaded with a CacheLocation
Lazily load Caches into a deep NetworkNode's CacheLocation rather than
setting its CacheLocation and greedily loading Caches into it at the
time of NetworkNode.generateTree(). This allows changes in Cache status
to be reflected in deep CZF lookups (i.e. when a Cache is set to
OFFLINE, stop deep routing to it). Whenever a new CRConfig is processed,
clear out deep Caches from the CacheLocations so that they can be lazily
loaded again (similar to how non-deep CacheLocations are reset to null).