Added missing remap lines for cachekey to mids.
The mids had logic to generate cacheurl parameters, but not cachekey
parameters. This was causing delivery services to effectively
invalidate at the mid when they were switched from cacheurl to
cachekey.

This adds in the logic for cachekey values at the mid in exactly the
same way it works for the edges.