Move binaries to /usr/sbin to conform to LFH, LSB
The Linux Filesystem Hierarchy standard is somewhat ambiguous about
where services go, and what constitutes a system vs user app. But
the community consensus seems to be /usr/sbin.

Irrespective, The Linux Foundation has made it clear /opt is not
preferred, and is for projects who "don't play well with others".

The Linux Filesystem Hierarchy is referred to by the Linux Standard
Base, and its locations (/etc, /var/log, /usr/bin, /usr/sbin) are the
preferred organisation for Linux projects.