Change TO Dockerfile RPM to omit version, URI
The original intent was for the Dockerfile to use the latest published
RPM. It seems clear that isn't working, though, and that the
Dockerfile needs versioned with the rest of Traffic Control. Users
wishing to deploy a specific release version should check out the
Dockerfile from that git revision. This changes the Dockerfile's
default RPM to not have any version, and not pull from the website.
The common usage should be to pass in an RPM built from the same
git revision as the Dockerfile.