WIP Add TM astats-dstypes stats format
Adds a Monitor stats format similar to Astats, but with FQDNs replaced
by Delivery Service names (xml_id). This is the ideal, but Apache
Traffic Server is not easily capable of it. However, some community
members have expressed a desire for it, and it seems reasonable to
provide the plugin with TC.

WIP - I haven't tested this yet; issues are unlikely, as it's a simple
modification, but it still needs tested before merging.