Fix steering race condition in TR
Don't proactively remove Steering data for delivery services not yet in
the CRConfig.json. This data is not "invalid" and can safely be kept.
This fixes the race condition where a user updates steering assignments
for delivery services that don't exist in CRConfig yet that causes the
steering delivery service to be unusable.