Fix subtle bug sending 304 without if-modif hdr
Fixes a subtle bug wherein under certain cirumstances a 304 with no
body was returned, when the user had NOT sent an `if-modified-since`.

This specifically happened when the object was in cache, and was
stale, but was permitted to be revalidated.

The bug was a subtle logic ordering error in the RetryingGetter.
Specifically, it checked if it could cache the response (which it
couldn't a standard 304 with no Cache-Control), _before_ checking if
the response was a 304 as a result of a revalidate.

This fixes the ordering.

Also fixes the revalidate check to only check for 304, not other 3xx
codes, as 304 is the only response permitting reuse.

Also moves the if-modified-since header setting to the retryinggetter,
out of the main handler, to reduce duplication, and put it in the
properly responsible location.