Add ifmodifiedsince client req 304, fix ims bugs
Adds an If-Modified-Since beforerespond plugin, to return a 304 to
clients who send an IMS header if the object is unmodified.

Also fixes some IMS-related bugs. E.g. a client IMS header is
stripped before requesting from the parent. Before, this was sent,
the 304 was cached and erroneously returned to future non-IMS reqs.