Fix transport to use correct proxy per rule
Fixes an issue with the shared transport pointer using the proxy from
previous requests, and racing as to which last-used proxy was used.

This fixes the transport to be shared for each to-parent, for
efficiency (as the Go docs suggest), but unique for different to
rules, and therefore not racing and overriding different proxies
for different rules.