Fix getter bug returning wrong object
There was a getter bug, returning the wrong object, specifically from
a different concurrent request. This fixes it.

It also rewrites the entire getter to be simpler. Now, the getter only
uses the existing goroutine, doesn't create more, and the data passing
to share the fetched object is far simpler. See the comments in the
thread/getter.go file for a detailed explanation of the new algorithm.