Fix HTTP/2
Fixes HTTP/2, by calling the http2.ConfigureServer func, and
continuing to set h2 in tls.Config.NextProtos.

Also changes InterceptListener to wrap the net.Conn not tls.Conn,
which breaks H2 because it casts to tls.Conn and explodes violently
if that fails.

Note this demonstrably, permanently breaks connection byte counting
for HTTP/2 connections. We still have the less-accurate
ResponseWriter.Write count for stats. But, the Go language simply
doesn't support per-request connection data, because H2 streams are
interleaved on the same real TCP connection, and the Go libraries
don't do the extra work to associate frame bytes with streams and
handlers.