Add URI Signing config files automatically when necessary.
When URI Signing is enabled on a Delivery Service assigned to a server,
that server will need to have the appropriate URI Signing config files
requested so that is has the appropriate keys and configuration.
Typically, this is done with profile parameters, but because the need
for the parameter varies with the configuration of the DS assigned to
the server, it's tedious and error prone to manage.

This change will make Traffic Ops automatically add the config file to
the list of required files for any server that requires it and no
servers that don't. It should be seamless for URI Signing users. If a TO
user needs to override the location of a URI Signing config file for
some reason, they can provide a location parameter and it will be
respected.

This does incur an additional hit on the database to collect the list of
URI Signed DSs, but it doesn't appear to impact performance beyond the
level of statistical noise.

Closes #1662