Add information to errors when cdns/routing cannot contact TR.
When a TR cannot be contacted, the cdns/routing endpoint returns a 500.
However, without information about why the 500 is being thrown, it can
be tricky to track down precisely what is going wrong. This adds the
name of the uncontactable host to the text of the error return as well
as adding a separate log entry, at ERROR level, to indicate the issue.