Protect ORT from errors caused by missing data in TO.
If a required file is not configured with a location parameter, ORT is
still attempting to process it, despite the fact that there is no file
name to process. This displays lots of errors that don't clearly
indicate what is happening or to which file it is happening. This error
message documents the problematic file and prevents its processing so
that things can continue, just without that file.