Add a Location API
Adding a Location API is the first step to refactoring lat/lon out of
the Cache Group API into its own entity - Location. With the Location
API in place, the cachegroup table can be updated to reference a
Location, and other future entities (such as Origins) can make use of
Locations as well.

This is part of the larger "geolocation-based client steering" effort.