Add CRConfig DS Modified fields
Adds a "modified" field to the CRConfig deliveryServices objects,
indicating the last time that DS's database table row was modified.

Adds a "anyModified" field to the CRConfig deliveryServices objects,
indicating the last time any database field used to create the JSON
object was modified, also including the server assignments (which
are in the contentServers key, not the deliveryServices key)

Adds a "version" field to the CRConfig object, indicating the API
version used to create that CRConfig.

Adds database triggers updating the deliveryservices last_modified
timestamp when a ds_regex, ds_server, or staticdnsentry is deleted.

These features are being added, with the goal of making it possible
for Traffic Router to determine if any individual DS in the CRConfig
has been modified, and only reload modified DSes, not everything.