Fix TO Go profileparameters to match Perl
Fixes the /profileparameters endpoint to match the old Perl. This
specifically matches Perl code, not the docs, because documentation
doesn't exist for this endpoint. I also wasn't able to run the Perl,
it appears to time out and returns a 502; but this change reflects
the Github case, and what the Perl code appears to do.

Fixes #2556