Added missing struct fields in client to satisfy v11 defintion of a user
Removed create and delete servers in the tests since the users test
doesn't need it, and the server tests were failing at the time I wrote
this.

I added a small test to verify that the correct struct fields existed,
looking at the struct fields in the localhost is insufficient since
the only place the `User` struct is used is in the client.

Note: The struct I modified will eventually be deprecated because it
isn't nullable, but in the meantime a User should still contain all
of its fields.