Fixed some issues with building CIAB
Makefile was unable to operate with '-j' because nothing was creating a Tomcat RPM, and would fail on systems that had an 'rpm' binary but weren't 'el7'. Both fixed.
Also, the cache Docker build was broken because a new Trafficserver RPM replaced the old 7.1.4, but had a different name. Fixed.