Prevent Perl input record separator change in TO dbdump API
Due to the way this code was previously written, the first call to the
dbdump API is successful, but subsequent calls are broken because the
chomp() function no longer strips trailing whitespace from the output of
the `hostname` command. This ends up breaking the response headers
because of the unexpected newline, causing traffic_ops_golang to emit
this error and return a 502:

    http: proxy error: net/http: HTTP/1.x transport connection broken:
    malformed MIME header line: -20181031125221.pg_dump"

By only locally declaring $/ (undefined), subsequent calls to `chomp()`
actually strip newlines because `$/ == "\n"` by default.