Adding backup file support for the crconfig and tmconfig files. Any time TM receives a valid cr or tm config file it gets written to a backup file on disk. A poller is started with a default of 1 sec (or using file_retry_timeout_ms in the config file) that will poll the TO url every interval. If the url is unreachable and the backup files exist then TM will fall back to using those backup files for monitoring. The url will then be checked on the next interval and the next time TO is available TM will switch back to the url usage.
Along with the file_retry_timeout_ms this also adds crconfig_backup_file and tmconfig_backup_file optional settings to specify backup file locations

On startup we continue to use the new backoff method however after we have reached TrafficOpsFileFallbackRetries amount we then test if the backup files exist and if they do we proceed to the disk based method. This way if TM starts up and TO
is down it will continue to work as long as it has a valid backup cr and tm config files.