Change TM stat history to sync.Map
This significantly improves performance. In my testing, for a large
CDN, processing Stat results can take ~5 seconds, where ~1s of that
was copying and setting the mutexed stat history variable. This
essentially eliminates that 1s or 1/5 of stat processing time, for
every stat process. This significantly reduces CPU, and/or makes
the monitor able to poll stats in significantly less time.

Also changes the stat history array to not reallocate memory,
but move the data when prepending a new result. Which is also
significantly faster, in my testing.