Fix TO Go auth wrapper to be after other wrappers
This specifically fixes a major bug, where auth failure returns a 200.

This causes a major bug in the TO Go client, (and likely other
clients)which assumes a 200 to be success, and then goes on to request
data, gets more auth failures with 200's, but thinks they're success,
and thinks the JSON is just missing all fields, which are defaulted
to 0/empty string/null.

This causes other bugs that just went unnoticed, like missing the
Access-Control headers, X-Server-Name, and gzipping.