Fix ordering of DS matchlists to match the Perl (#3066) (#3068)
The DS matchlist needs to be ordered by the set_number just like it was
done in the Perl. There are clients that depend on that specific
ordering, e.g. Traffic Portal which uses the first exampleURL to create
the hostName field of an "add cert to DS" request.

(cherry picked from commit 53e9e321bdc5d3a8139ef283630b75e7cf352a92)