Fix TO CRConfig generation to use tm.url parameter
Also adds a config param, so it can be set back without a new binary
deployment, if someone discovers it needs to be the host header for
some reason, default false.

Also changes the CRConfig stats.tmpath to be the real path, and adds
a config param to set it back if there's an issue, default false.

Fixes #2224