Fix TO Go wrapper copying req, breaking codes
Fixed the Traffic Ops Golang Auth wrapper to not copy the Request.

Wrappers should never copy the Request, because it breaks the context,
set by other wrappers or the real handler (really, the whole reason
the context exists).

Right now, it was specifically breaking the response code context
key; but in the future, it could break anything. Wrappers should
never change the *http.Request pointer, always the value. That is,
`*req = *foo`, not `req = foo`.