Fix improper file slurping in Perl
The proper way to enable slurp mode is by doing `local $/;`. When you do
`undef $/` it is undefining the global `$/` input record separator which
is bad and can cause a multitude of problems. Using the `local` version
only nullifies the input record separator locally in order to enable
slurp mode in that particular function rather than changing it globally.