Fixed routes for users
The regex for the get was written to combine the id and non-id case,
then incorrectly copied for other user endpoints. It incorrectly allowed
a non-id case for PUT, and the commented out DELETE.

The regex is updated so that the id and non-id case are two separate
regexes. They were separate before, so the GET method had a redundant
regex until now.