Fix TO cg/ds failure when params exist
Fixes cachegroups/id/deliveryservices failing when one of the location
parameters being created on the newly-assigned servers already exists
(because a different server on the same profile was previously
assigned to the delivery service).

This causes the query "ON CONFLICT DO UPDATE SET id = EXCLUDED.id
RETURNING id" to fail, because changing the id creates a foreign
key conflict on the profile_parameters table.

The query doesn't actually need to update the id, Postgres just
requires _something_ be updated in order to return. This changes it to
set name=name, which is the same, and doesn't cause a query failure
from a fk conflict.