Fix nil pointer dereference in PUT /deliveryserivces/{id} endpoint
If this endpoint receives a json request without a "matchList" array
(which is not a required field), the endpoint will panic on a nil
pointer dereference.

Load the delivery service matchList from the DB before dereferencing it.
As a side effect, this also prevents a user from passing in their own
"matchList" array which would potentially cause the wrong hostname to be
updated into the delivery service's ssl keys.