CIAB: Fix Traffic Vault Enrollment
Since we need to follow the services isolation policy, it should not set dependency in docker-compose.yml.
So if TO-perl run faster than TV, the script `set-to-ips-from-dns.sh` can not obtain TV's IP and supplly it to `040-trafficvault_server.json`

This commit fix TV enrollment so TV can register by itself.