Records.config ds profile parameters (#3279)
* Allow DS profile to create records config overrides

These changes allow you to put records.config parameters in to a DS profile and have them appear as remap overrides.  This is similar to how cachekey works in the DS profile. If the file given is for records.config then
options are generated for the remaps from the DS in this manner:
@plugin=conf_remap.so @pparam=param_name=value

This way you can do parameter overrides without having to use raw remap.

The parameters are expected to match the existing usages of records.config parameters in the scheme of:
CONFIG paramater_name INT/STRING/FLOAT value

outside of this the parameter will be rejected

* Change param splitting to only 2 entries, allows for strings with spaces in case for some strange reason thats a thing