Bugfixes for users endpoint (#3218)
* Fixed routes for users

The regex for the get was written to combine the id and non-id case,
then incorrectly copied for other user endpoints. It incorrectly allowed
a non-id case for PUT, and the commented out DELETE.

The regex is updated so that the id and non-id case are two separate
regexes. They were separate before, so the GET method had a redundant
regex until now.

* Fixed bugs related to privilege, role checking, then added integration tests for them.
Added role to CurrentUser struct

* Prevent defer from hanging in case of error

* Updated docs for users endpoint

Removed PUT users documentation
Removed deprecation of PUT users/:id
Removed messages referencing github issues that will are/will be resolved

* Fix rolename regression

* fixup lying comments

* Responded to PR comments

* Added api test for rolename regression

* Removed accidental fmt

* minor stuff

* Fixed rolename regression in a different way

Removes duplicated data in struct and keeps any duplicate struct in just the server-side.

* Added api test to check the capitalization of rolename.

Unlike before, the json marshaller isn't used because the client user struct only supports one capitalization of rolename, and golang's json marshaller is case-insensitive, which wouldn't give the necessary information.

* Minor documentation updates

* Removed unnecessary routes and fixed test

* Changed user endpoints to be v1.1 instead of v1.4 so they can effectively replace the perl