Revert "Use JSONIntStr for user endpoint"
This reverts commit ccec233e2c7951e28b42bf370e5fc6b5c498df95.

In the conversion from perl to go, perl will accept strings or ints as
integers because it is a weakly typed language. The use of JSONIntStr
allowed the same to happen in golang (in this case for a single
endpoint).

The use of JSONIntStr would keep the functionaility from perl, however I
am reverting the change because keeping this behavior implies that we
want to maintain it in the future for all other endpoints.

When the original commit was made, there was not yet a consensus to keep the bug as is.
Here is a part of the discussion (the rest was in slack):
https://lists.apache.org/thread.html/f0624d4386f1de1f18b322436df6fa0402bf34314abd4b108edc5fde@%3Cdev.trafficcontrol.apache.org%3E