Fix DS Snapshots where qry to be outside times
The where query part was previously inside the last_updated sorting.
It has to be outside, because the where filtering needs to happen
after getting the latest snapshot row.

For example, if a DS is changed to inactive, putting the where
inside the 'where last_updated < snap order by last_updated'
will make the latest inactive row be excluded, and the latest
active row selected.