Made configuration changes for CIAB.
Added variable "ADD_DEMO1_SSLKEYS", which determines whether traffic ops
is configured with riak and whether the ssl key demo (requiring riak) is
tried. Perhaps it would be ideal to auto-detect whether traffic vault is
running to replace either the first conditional or both, but I'm not sure
that's possible.