Fixed xmlids and routing names not allowed to start with numeric characters
Also changed around some button types; 'submit' is the default and
will cause the user-agent to validate the form on click. This was
happening with the "show/hide advanced" button which is now
`type="button"` to prevent this. The save/fulfill buttons are now
`type="submit"` because that's when form validation ought to be
done.