Added ScanIntoItems and ScanIntoItem functions
First created a ScanIntoItems function that basically unpacks a list
into sql's scan function.

Then updated the generic cruder to use the function, although it is meant
more for endpoints that don't already have a generic function. I updated
the generic case since it could be used as a model and so that the error
messages won't look too different from implementors that use
ScanIntoItems.

ScanIntoItems is especially convienent when returning a number of items. #2494 is
related.

Here is an example that is applicable for cachegroups:
```
items := []interface{}{
    &cg.ID,
    &cg.ParentName,
    &cg.SecondaryParentName,
}
if err = api.ScanIntoItems(resultRows, items); err != nil {
    if err == sql.ErrNoRows {
	return errors.New("no cachegroup found with this id"), nil, http.StatusNotFound
    }
    return nil, err, http.StatusInternalServerError
}
```

Mayber later down the line, ScanIntoItems would be prefered over
ScanStruct since ScanStruct belongs to sqlx. I don't think we use sqlx
for anything other than ScanStruct, so we could potentially get rid of
a dependency.