Fixed user cards taking forever to fetch role names
Previously, each card made its own HTTP request to get a name for its Role.
The new implementation makes only 1 request at the users page level to
fetch a map of role IDs to role names. This behaviour is optional when using
a user-card, and it will fall back to the old behaviour if no map is passed.