Now 100% more configurable at run time!
The server executable now offers options to change the port on which
it listens, the TO server information, and certificate verification
(as well as standard version and help options). Also, the server
will now check connectivity with the specified TO server at startup
(via `/api/1.4/ping`), and will exit with an error if the check fails.