Added IDImpl struct
IDImpl is a struct that implements GetKeys and SetKeys for the common
case where the only key is a single numerical ID. The idea is to make
a single, common implementation that can be used across all endpoints
that can use it. As a bonus these endpoints don't need to make a GetKeys
or SetKeys function, nor do they have to provide the json and db tags
and we know the behavior we want is implemented.

The downside is that current swagger comments may be broken (although we
weren't exactly using it anyway).

This is mostly a test to get feedback. I think it is relatively useful,
although it is debatable whether or not this works well as a layer of
abstraction since "every abstraction layer has a cost". Also, some might
perfer to keep the tc structs 'clean'.

For now, I've only made the CDN struct use the IDImpl. IDImpl only works
for Nullable structs (the structs that we define endpoints on).