Fix TO log format for 3rd party libs
Fixes TO Go logging of Riak and ReverseProxy, which weren't properly
adding the prefix and date format of lib/go-log.

Fixes it by adding a func to create a new wrapper Logger, which
writes to the given logger with the standard log.Printf funcs,
properly setting the time and prefix.

Fixes #3587