Fix TM Thresholds
This fixes the Monitor racing and the health poll marking Available
caches which were marked Unavailable by the Stat poll.

There was already logic to prevent that, but it wasn't accounting
for the Calculated stats, so the Health poll didn't check Calculated
stats for thresholds, but it does "have" those stats, so it thought
it could mark them available again.

This fixes the Health poll to also check thresholds for Calculated
stats, which not only fixes the race, but makes thresholds be
marked up and down faster with the quicker Health poll.