Traffic Ops Golang parent.config (#3075)
* Add TO client ATS config funcs

* Fix missing Apache header

* WIP Add TOGo parent.config

* WIP TO Go parent.config fix mid details

* Fix TO Go parent.config spacing to match Perl

* Add TO parent.config API test

* Fix TO parent.config bad round_robin default

parent.config was defaulting to 'round_robin=urlhash', which is not
a valid ATS config. This fixes it to default to consistent_hash,
the usual default.

Fixes #3071

* Fix TO parent.config to sort edge parents by rank

Fixes #2725

* Add doc comment to to.session request, rawRequest

* Change TO API Test to use With helper

* Remove debug prints, fix parent.config typo

* Change TOGo parent.config to API 1.1+

* Remove TO unnecessary comment

* Remove TO TODO to determine whether to warn

Remove comment asking if a warning should be logged when a DS has
no parameters, per PR Review.

I am unsure, but I believe this to be normal behavior and not require
a warning.

* Remove TO unnecessary comments

* Renamed TO CGPDeliveryService to CGServer

More accurately reflects what it is and the data it contains. It's
not a great name, but I'm not sure what would be better.

As requested in PR Review.

* Change TO parentdotconfig to use type enums

* Renamed TO variable

As requested in PR Review.

* Rename TO parentdotconfig symbols

As requested in PR Review.

* Fix TO parent.config missing defaults

* Move TO RemoveStrDuplicates to lib/go-util

Also renames local variables, per PR Review.

* Rename TO parent.config variable

Per PR Review.

* Change TO comment to TODO

* Fix TO error log typo

* Add TO TODO for improving parent.config logic

* Change TO parent.config var to longer name

Per PR Review